GeoPin/pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "geopin"
version = "0.1.0"
description = "A tiny open-source location visualization tool for AI agents."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
  {name = "Open source contributors"}
]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.35.0",
  "httpx>=0.28.0",
  "mcp[cli]>=1.2.0",
  "pydantic>=2.10.0",
  "pywebview>=6.0.0"
]
keywords = ["mcp", "agent", "maps", "globe", "openclaw", "hermes"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]

[project.urls]
Homepage = "https://github.com/yourname/geopin"
Repository = "https://github.com/yourname/geopin"
Issues = "https://github.com/yourname/geopin/issues"

[project.scripts]
geopin = "geopin.cli:main"
geopin-mcp = "geopin.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/geopin"]

[tool.hatch.build.targets.sdist]
include = [
  "/src",
  "/README.md",
  "/LICENSE",
  "/docs",
  "/adapters",
  "/examples"
]

[tool.hatch.build.targets.wheel.force-include]
"src/geopin/static" = "geopin/static"
